+++ contents/ui/configAbout.qml
@@ -37,8 +37,12 @@ KCM.SimpleKCM {
     property var cfg_enabledPartitionsDefault
     property var cfg_partitionOrder
     property var cfg_partitionOrderDefault
+    property var cfg_partitionOptOut
+    property var cfg_partitionOptOutDefault
     property var cfg_partitionLabels
     property var cfg_partitionLabelsDefault
+    property var cfg_diskPartitionColors
+    property var cfg_diskPartitionColorsDefault
     property var cfg_showCpuCores
     property var cfg_showCpuCoresDefault
     property var cfg_mergeCpuTemp